Move LetterPrintStatus
